params.pyの型修正

+++ src/optimize/params.py
@@ -26,8 +26,8 @@ class ArtificialDataParameter:
     num_of_other_features: int
     depth_of_trees: int
     base_price: int
-    price_min: int = 4
-    price_max: int = 6
+    price_min: float = 4.0
+    price_max: float = 6.0
     base_quantity: int = 300
     seed: int = 0
     data_type: str = "artificial"
